% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compact.R
\name{adjustForCompaction}
\alias{adjustForCompaction}
\title{Adjust for compaction based on parameters}
\usage{
adjustForCompaction(origDepth, maxCompact, compactOver, eEnd = 3)
}
\arguments{
\item{origDepth}{unadjusted depth in cm. Can be negative so long as it's > -maxCompact}

\item{maxCompact}{Maximum amount of compaction in cm (i.e., the compaction observed at the surface)}

\item{compactOver}{Depth over which to compensate in the compaction}

\item{eEnd}{Where to end the exp function scale. Here, a parameter that controls the linearity of the the compaction. Lower values are more linear. Must be greater than 0.  (Default =3)}
}
\value{
adjusted depth
}
\description{
Adjust for compaction based on parameters
}
\examples{
depths <- seq(0,6,by = .1)
ad <- purrr::map_dbl(depths, adjustForCompaction, maxCompact = 0.5, compactOver = 5)

}
