% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreSectionDepth.R
\name{dblf_to_coreSection}
\alias{dblf_to_coreSection}
\title{Get core section depth from dblf}
\usage{
dblf_to_coreSection(core, dblf)
}
\arguments{
\item{core}{Name of the core (not the core section!) (e.g. "L380_OTOTO_LC4U")}

\item{dblf}{depth below the lake floor in cm}
}
\value{
a data.frame with core section and depth
}
\description{
Get core section depth from dblf
}
\examples{
dblf_to_coreSection("L380_OTOTO_LC4U",110)
dblf_to_coreSection("L380_OTOTO_LC1U",10)
}
