% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreSectionDepth.R
\name{findCoreSectionName}
\alias{findCoreSectionName}
\title{Find a core section name}
\usage{
findCoreSectionName(corename_guess)
}
\arguments{
\item{corename_guess}{}
}
\value{
Potential matches of core names given a guess
}
\description{
Find a core section name
}
\examples{
findCoreSectionName("ngano")
}
