% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreSectionDepth.R
\name{hsi_to_dblf}
\alias{hsi_to_dblf}
\title{Convert depths from hyperspectral imaging}
\usage{
hsi_to_dblf(corename, cm, extraAllowedBottom = 1)
}
\arguments{
\item{corename}{a vector of corenames}

\item{cm}{a vector of corresponding depths that matches the length of corenames}

\item{extraAllowedBottom}{how many cm below the bottom of the ROI are allowed (converted to bottom of ROI) (default = 1)}
}
\value{
depth below lake floor in cm
}
\description{
Convert depths from hyperspectral imaging
}
\examples{

hsi_to_dblf("L380_DUNCA3_LC4U_2",0:30)

}
