% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreSectionDepth.R
\name{listCoreSectionNames}
\alias{listCoreSectionNames}
\title{List known core names}
\usage{
listCoreSectionNames()
}
\value{
a list of known core names
}
\description{
List known core names
}
\examples{
allNames <- listCoreSectionNames()
}
