% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreSectionDepth.R
\name{coreSection_to_dblf}
\alias{coreSection_to_dblf}
\title{Calculate depth below lake floor}
\usage{
coreSection_to_dblf(corename, cm)
}
\arguments{
\item{corename}{Name of the lakes 380 core section}

\item{cm}{section depth, from the top of the core liner, in cm}
}
\value{
depth below lake floor in cm
}
\description{
Calculate depth below lake floor
}
\examples{

coreSection_to_dblf("L380_DUNCA3_LC4U_1",50:60)

}
