% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{dblf_from_file}
\alias{dblf_from_file}
\title{Title}
\usage{
dblf_from_file(filename = NA, outfile = NA, conv.type = "coreliner")
}
\arguments{
\item{filename}{Optionally enter the file name (default = NA, will let you pick)}

\item{outfile}{Optionally enter the output file path and name (default = NA, will append "-dblf" and save as csv}

\item{conv.type}{which conversion to use. For now the only option is coreliner}
}
\value{
an updated file
}
\description{
Title
}
